Remove doubled stats issue (zenflow 7d6d4f80)
We have a conceptual issue in our data thats affect all stats calculations in our players and team statistic. The issue is related to traded players, the latest example is Lamar Jackson traded from BAL to CLE, when im lookung into Cle stats i see Lamar is there from week 1, but actual trade happened at week 5, so we need to find a way to properly remove him and any other traded player from our calculations.

This is not simple, we need to think a lot how to do so, look at every CSV to understand the best way. Before the implementation and any coding suggest the ideas for review